+++ packages/stk/stk_mesh/stk_mesh/base/Bucket.cpp
@@ -272,11 +272,9 @@ Bucket::Bucket(BulkData & mesh,
     m_node_kind(INVALID_CONNECTIVITY_TYPE),
     m_edge_kind(INVALID_CONNECTIVITY_TYPE),
     m_face_kind(INVALID_CONNECTIVITY_TYPE),
-    m_element_kind(INVALID_CONNECTIVITY_TYPE),
     m_fixed_node_connectivity(),
     m_fixed_edge_connectivity(),
     m_fixed_face_connectivity(),
-    m_fixed_element_connectivity(),
     m_dynamic_node_connectivity(initialCapacity),
     m_dynamic_edge_connectivity(initialCapacity, should_store_permutations(entityRank, stk::topology::EDGE_RANK)),
     m_dynamic_face_connectivity(initialCapacity, should_store_permutations(entityRank, stk::topology::FACE_RANK)),
@@ -294,7 +292,6 @@ Bucket::Bucket(BulkData & mesh,
   setup_connectivity(m_topology, entityRank, stk::topology::NODE_RANK, m_node_kind, m_fixed_node_connectivity);
   setup_connectivity(m_topology, entityRank, stk::topology::EDGE_RANK, m_edge_kind, m_fixed_edge_connectivity);
   setup_connectivity(m_topology, entityRank, stk::topology::FACE_RANK, m_face_kind, m_fixed_face_connectivity);
-  setup_connectivity(m_topology, entityRank, stk::topology::ELEMENT_RANK, m_element_kind, m_fixed_element_connectivity);
 
   m_parts.reserve(m_key.size());
   supersets(m_parts);
@@ -313,7 +310,6 @@ size_t Bucket::memory_size_in_bytes() const
   bytes += m_fixed_node_connectivity.heap_memory_in_bytes();
   bytes += m_fixed_edge_connectivity.heap_memory_in_bytes();
   bytes += m_fixed_face_connectivity.heap_memory_in_bytes();
-  bytes += m_fixed_element_connectivity.heap_memory_in_bytes();
   bytes += m_dynamic_node_connectivity.heap_memory_in_bytes();
   bytes += m_dynamic_edge_connectivity.heap_memory_in_bytes();
   bytes += m_dynamic_face_connectivity.heap_memory_in_bytes();
@@ -343,10 +339,7 @@ void Bucket::change_existing_connectivity(unsigned bucket_ordinal, stk::mesh::En
 
 void Bucket::change_existing_permutation_for_connected_element(unsigned bucket_ordinal_of_lower_ranked_entity, unsigned elem_connectivity_ordinal, stk::mesh::Permutation permut)
 {
-    stk::mesh::Permutation *perms = m_element_kind == FIXED_CONNECTIVITY ?
-        m_fixed_element_connectivity.begin_permutations(bucket_ordinal_of_lower_ranked_entity)
-        :
-        m_dynamic_element_connectivity.begin_permutations(bucket_ordinal_of_lower_ranked_entity);
+    stk::mesh::Permutation *perms = m_dynamic_element_connectivity.begin_permutations(bucket_ordinal_of_lower_ranked_entity);
 
     if (perms)
     {
@@ -660,7 +653,7 @@ Bucket::grow_capacity()
   m_dynamic_other_connectivity.increase_bucket_capacity(m_capacity);
 }
 
-bool Bucket::destroy_relation(Entity e_from, Entity e_to, const RelationIdentifier local_id )
+bool Bucket::destroy_relation(Entity e_from, Entity e_to, const Ordinal local_id )
 {
   const unsigned from_bucket_ordinal = mesh().bucket_ordinal(e_from);
   DestroyRelationFunctor functor(from_bucket_ordinal, e_to, static_cast<ConnectivityOrdinal>(local_id));
@@ -820,9 +813,6 @@ void Bucket::debug_check_for_invalid_connectivity_request([[maybe_unused]] Conne
     }
     else if (type == &m_face_kind) {
       rank = stk::topology::FACE_RANK;
-    }
-    else if (type == &m_element_kind) {
-      rank = stk::topology::ELEMENT_RANK;
     }
     else {
       STK_ThrowAssert(false);